Remplacement de quelques int64_t par des int32_t
Remplacement des int64_t par des int32_t quand les variables représentent
des millimètres ou des degrés. Normalement, ça ne devrait pas avoir
d'impact (overflow), vu qu'on demande rarement au robot de parcourir +2000km
ou de faire près de six millions de tours sur lui même...

+++ asserv/Utils/Utils.h
@@ -13,10 +13,10 @@ public :
     int64_t static constrain(int64_t value , int64_t min , int64_t max);
 
     // Conversions
-    int64_t static mmToUO(Odometrie *odo, int64_t valeur);
-    int64_t static degToUO(Odometrie *odo, int64_t valeur);
+    int64_t static mmToUO(Odometrie *odo, int32_t valeur);
+    int64_t static degToUO(Odometrie *odo, int32_t valeur);
     int64_t static radToUO(Odometrie *odo, double valeur);
-    int64_t static UOTomm(Odometrie *odo, int64_t valeur);
+    int32_t static UOTomm(Odometrie *odo, int64_t valeur);
 
 };
